Move controls out into a component
